fast sampling

+++ model.py
@@ -14,26 +14,31 @@ import stan_utils as stan
 
 plt.style.use(mpl_style)
 
-def _savefig(fig, basename, **kwargs):
+def _savefig(fig, basename, dry_run=True, **kwargs):
     kwds = dict(dpi=300)
     kwds.update(kwargs)
-    fig.savefig("figures/{}.pdf".format(basename), **kwds)
-    fig.savefig("figures/{}.png".format(basename), **kwds)
+    if not dry_run:
+        fig.savefig("figures/{}.pdf".format(basename), **kwds)
+        fig.savefig("figures/{}.png".format(basename), **kwds)
+    else:
+        print("Dry-run: not saving figure")
 
-"""
-gaia_sources = Table.read("data/rv_floor_cal-result.fits")
+    return None
 
-# Take a random subset of stars.
-subset_indices = np.random.choice(len(gaia_sources), size=1000, replace=False)
-subset = gaia_sources[subset_indices]
-"""
+
+#gaia_sources = Table.read("data/rv_floor_cal-result.fits")
 
 subset = Table.read("data/rv_floor_cal_subset-result.fits.gz")
+
+# Take a random subset of stars.
+subset_indices = np.random.choice(len(subset), size=1000, replace=False)
+subset = subset[subset_indices]
 
 
 rv_variance = np.array(subset["radial_velocity_error"])**2
 apparent_magnitude = np.array(subset["phot_g_mean_mag"])
 flux = np.array(subset["phot_g_mean_flux"])
+rp_flux = np.array(subset["phot_rp_mean_flux"])
 
 scatter_kwds = dict(s=1, facecolor="#000000", alpha=0.05, rasterized=True)
 
@@ -113,7 +118,80 @@ _savefig(fig, "dumb-initial-fit")
 
 
 # OK, let's prepare a model.
+qc = np.isfinite(rp_flux)
+N = sum(qc)
+data = dict(N=N, rv_variance=rv_variance[qc], flux=rp_flux[qc])
+
+mu_design_matrix = np.array([
+    np.ones(N),
+    data["flux"]**-1,
+    data["flux"]**-2,
+])
+sigma_design_matrix = np.array([
+    np.ones(N),
+    data["flux"]**-1,
+    data["flux"]**-2,
+])
+data.update(
+    mu_design_matrix=mu_design_matrix.T, 
+    sigma_design_matrix=sigma_design_matrix.T,
+    M=mu_design_matrix.shape[0],
+    S=sigma_design_matrix.shape[0])
+
+init = dict(outlier_fraction=0.5,
+    mu_coefficients=np.array([1e10, 1e5, 1e-2]),
+    sigma_coefficients=np.array([1e10, 1e5, 1e-2]))
+
+
 model = stan.load_stan_model("model.stan")
+p_opt = model.optimizing(data=data, init=init)
+
+
+fig, ax = plt.subplots()
+ax.scatter(data["flux"], data["rv_variance"], **scatter_kwds)
+ax.plot(xi, np.polyval(p_opt["mu_coefficients"], 1.0/xi), c="r", lw=2, zorder=1000)
+ax.semilogx()
+ax.set_xlabel(r"\textrm{mean rp flux}")
+ax.set_ylabel(r"\textrm{radial velocity variance} $(\textrm{km\,s}^{-1})$")
+ax.set_xlim(1.25e4, 1e8)
+ax.set_ylim(-0.1, 1.5)
+
+fig.tight_layout()
+#_savefig(fig, "stan-initial-fit")
+
+
+
+samples = model.sampling(**stan.sampling_kwds(data=data, chains=2, iter=2000,
+    init=init))
+
+
+raise a
+
+#plt.style.use({"text.usetex": False })
+#fig = stan.plots.traceplot(samples, ("outlier_fraction", "c0", "s0"))
+#fig.tight_layout()
+#_savefig(fig, "stan-trace")
+
+c0 = samples.extract("c0")["c0"].mean()
+c1 = samples.extract("c1")["c1"].mean()
+
+fig, ax = plt.subplots()
+ax.scatter(data["flux"], data["rv_variance"], **scatter_kwds)
+ax.plot(xi, c0 + c1 * xi**-2, c="r", lw=2, zorder=1000)
+ax.semilogx()
+ax.set_xlabel(r"\textrm{mean rp flux}")
+ax.set_ylabel(r"\textrm{radial velocity variance} $(\textrm{km\,s}^{-1})$")
+#ax.set_xlim(1.25e4, 1e8)
+#ax.set_ylim(-0.1, 1.5)
+
+s0 = samples.extract("s0")["s0"].mean()
+s1 = samples.extract("s1")["s1"].mean()
+
+fig, ax = plt.subplots()
+ax.plot(xi, s0 + s1 * xi**-2, c="r", lw=2)
+ax.semilogx()
+ax.set_xlabel(r"\textrm{mean rp flux}")
+ax.set_ylabel(r"\textrm{scatter}")
 
 raise a
